+++ src/main/java/com/dalo/patterns/creationalPatterns/abstractFactory/dao/ClientDao.java
@@ -1,0 +1,7 @@
+package com.dalo.patterns.creationalPatterns.abstractFactory.dao;
+
+import com.dalo.patterns.creationalPatterns.abstractFactory.entity.Employee;
+
+public interface ClientDao {
+    Employee getClientById(Long id);
+}
