+++ src/main/java/com/dalo/patterns/creationalPatterns/abstractFactory/dao/EmployeeDao.java
@@ -1,0 +1,7 @@
+package com.dalo.patterns.creationalPatterns.abstractFactory.dao;
+
+import com.dalo.patterns.creationalPatterns.abstractFactory.entity.Employee;
+
+public interface EmployeeDao {
+    Employee getEmployeeById(Long id);
+}
