+++ src/main/java/com/dalo/patterns/creationalPatterns/abstractFactory/dao/hashtableImpl/HashTableClientDao.java
@@ -1,0 +1,12 @@
+package com.dalo.patterns.creationalPatterns.abstractFactory.dao.hashtableImpl;
+
+import com.dalo.patterns.creationalPatterns.abstractFactory.dao.ClientDao;
+import com.dalo.patterns.creationalPatterns.abstractFactory.entity.Employee;
+
+public class HashTableClientDao implements ClientDao {
+
+    @Override
+    public Employee getClientById(Long id) {
+        return null;
+    }
+}
