+++ src/main/java/com/dalo/patterns/creationalPatterns/abstractFactory/dao/hashtableImpl/HashTableEmployeeDao.java
@@ -1,0 +1,11 @@
+package com.dalo.patterns.creationalPatterns.abstractFactory.dao.hashtableImpl;
+
+import com.dalo.patterns.creationalPatterns.abstractFactory.dao.EmployeeDao;
+import com.dalo.patterns.creationalPatterns.abstractFactory.entity.Employee;
+
+public class HashTableEmployeeDao implements EmployeeDao {
+    @Override
+    public Employee getEmployeeById(Long id) {
+        return null;
+    }
+}
