+++ src/main/java/com/dalo/patterns/creationalPatterns/abstractFactory/dao/postgresImpl/PostgresClientDao.java
@@ -1,0 +1,11 @@
+package com.dalo.patterns.creationalPatterns.abstractFactory.dao.postgresImpl;
+
+import com.dalo.patterns.creationalPatterns.abstractFactory.dao.ClientDao;
+import com.dalo.patterns.creationalPatterns.abstractFactory.entity.Employee;
+
+public class PostgresClientDao implements ClientDao {
+    @Override
+    public Employee getClientById(Long id) {
+        return null;
+    }
+}
