+++ src/main/java/com/dalo/patterns/creationalPatterns/abstractFactory/entity/Client.java
@@ -1,0 +1,5 @@
+package com.dalo.patterns.creationalPatterns.abstractFactory.entity;
+
+public class Client {
+    private Long id;
+}
