+++ src/main/java/com/dalo/patterns/creationalPatterns/abstractFactory/entity/Employee.java
@@ -1,0 +1,5 @@
+package com.dalo.patterns.creationalPatterns.abstractFactory.entity;
+
+public class Employee {
+    private Long id;
+}
