+++ src/main/java/com/dalo/patterns/creationalPatterns/builder/model/Employee.java
@@ -1,0 +1,40 @@
+package com.dalo.patterns.creationalPatterns.builder.model;
+
+public class Employee {
+    private Long id;
+    private String firstName;
+    private String lastName;
+    private String phoneNumber;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public String getPhoneNumber() {
+        return phoneNumber;
+    }
+
+    public void setPhoneNumber(String phoneNumber) {
+        this.phoneNumber = phoneNumber;
+    }
+}
