+++ src/main/java/com/dalo/patterns/creationalPatterns/factoryMethod/dao/EmployeeDao.java
@@ -1,0 +1,8 @@
+package com.dalo.patterns.creationalPatterns.factoryMethod.dao;
+
+import com.dalo.patterns.creationalPatterns.factoryMethod.entity.Employee;
+
+public interface EmployeeDao {
+    Employee getEmployeeById(Long id);
+    void deleteEmployee(Long id);
+}
