+++ src/main/java/com/dalo/patterns/creationalPatterns/factoryMethod/dao/HashTableEmployeeDao.java
@@ -1,0 +1,15 @@
+package com.dalo.patterns.creationalPatterns.factoryMethod.dao;
+
+import com.dalo.patterns.creationalPatterns.factoryMethod.entity.Employee;
+
+public class HashTableEmployeeDao implements EmployeeDao {
+    @Override
+    public Employee getEmployeeById(Long id) {
+        return null;
+    }
+
+    @Override
+    public void deleteEmployee(Long id) {
+
+    }
+}
