+++ src/main/java/com/dalo/patterns/creationalPatterns/factoryMethod/factory/DaoFactory.java
@@ -1,0 +1,7 @@
+package com.dalo.patterns.creationalPatterns.factoryMethod.factory;
+
+import com.dalo.patterns.creationalPatterns.factoryMethod.dao.EmployeeDao;
+
+public abstract class DaoFactory {
+    public abstract EmployeeDao createEmployeeDao();
+}
