+++ src/main/java/com/dalo/patterns/creationalPatterns/singleton/dao/EmployeeDao.java
@@ -1,0 +1,7 @@
+package com.dalo.patterns.creationalPatterns.singleton.dao;
+
+import com.dalo.patterns.creationalPatterns.pototype.model.Employee;
+
+public interface EmployeeDao {
+    Employee getEmployeeByName(String name);
+}
