+++ src/main/java/com/dalo/patterns/creationalPatterns/singleton/dao/impl/EmployeeDaoImpl.java
@@ -1,0 +1,13 @@
+package com.dalo.patterns.creationalPatterns.singleton.dao.impl;
+
+import com.dalo.patterns.creationalPatterns.pototype.model.Employee;
+import com.dalo.patterns.creationalPatterns.singleton.dao.EmployeeDao;
+
+public class EmployeeDaoImpl implements EmployeeDao {
+
+    @Override
+    public Employee getEmployeeByName(String name) {
+        //implementation
+        return null;
+    }
+}
