+++ src/main/java/com/dalo/patterns/creationalPatterns/singleton/model/Employee.java
@@ -1,0 +1,4 @@
+package com.dalo.patterns.creationalPatterns.singleton.model;
+
+public class Employee {
+}
