+++ src/main/java/com/dalo/patterns/creationalPatterns/singleton/singleton/Singleton.java
@@ -1,0 +1,17 @@
+package com.dalo.patterns.creationalPatterns.singleton.singleton;
+
+public class Singleton {
+    private static Singleton instance;
+    private String value;
+    private Singleton(String value) {
+        this.value = value;
+    }
+
+    public static Singleton getInstance(String value) {
+        if (instance == null) {
+            instance = new Singleton(value);
+        }
+        return instance;
+    }
+
+}
