--- src/main/java/com/dalo/patterns/creational_patterns/abstract_factory/dao/ClientDao.java
@@ -1,7 +1,0 @@
-package com.dalo.patterns.creational_patterns.abstract_factory.dao;
-
-import com.dalo.patterns.creational_patterns.abstract_factory.entity.Employee;
-
-public interface ClientDao {
-    Employee getClientById(Long id);
-}
