--- src/main/java/com/dalo/patterns/creational_patterns/abstract_factory/dao/EmployeeDao.java
@@ -1,7 +1,0 @@
-package com.dalo.patterns.creational_patterns.abstract_factory.dao;
-
-import com.dalo.patterns.creational_patterns.abstract_factory.entity.Employee;
-
-public interface EmployeeDao {
-    Employee getEmployeeById(Long id);
-}
