--- src/main/java/com/dalo/patterns/creational_patterns/abstract_factory/dao/hashtable_impl/HashTableClientDao.java
@@ -1,12 +1,0 @@
-package com.dalo.patterns.creational_patterns.abstract_factory.dao.hashtable_impl;
-
-import com.dalo.patterns.creational_patterns.abstract_factory.dao.ClientDao;
-import com.dalo.patterns.creational_patterns.abstract_factory.entity.Employee;
-
-public class HashTableClientDao implements ClientDao {
-
-    @Override
-    public Employee getClientById(Long id) {
-        return null;
-    }
-}
