--- src/main/java/com/dalo/patterns/creational_patterns/abstract_factory/dao/hashtable_impl/HashTableEmployeeDao.java
@@ -1,11 +1,0 @@
-package com.dalo.patterns.creational_patterns.abstract_factory.dao.hashtable_impl;
-
-import com.dalo.patterns.creational_patterns.abstract_factory.dao.EmployeeDao;
-import com.dalo.patterns.creational_patterns.abstract_factory.entity.Employee;
-
-public class HashTableEmployeeDao implements EmployeeDao {
-    @Override
-    public Employee getEmployeeById(Long id) {
-        return null;
-    }
-}
