--- src/main/java/com/dalo/patterns/creational_patterns/abstract_factory/entity/Client.java
@@ -1,5 +1,0 @@
-package com.dalo.patterns.creational_patterns.abstract_factory.entity;
-
-public class Client {
-    private Long id;
-}
