--- src/main/java/com/dalo/patterns/creational_patterns/factory_method/dao/EmployeeDao.java
@@ -1,8 +1,0 @@
-package com.dalo.patterns.creational_patterns.factory_method.dao;
-
-import com.dalo.patterns.creational_patterns.factory_method.entity.Employee;
-
-public interface EmployeeDao {
-    Employee getEmployeeById(Long id);
-    void deleteEmployee(Long id);
-}
