--- src/main/java/com/dalo/patterns/creational_patterns/factory_method/dao/PostgresEmployeeDao.java
@@ -1,16 +1,0 @@
-package com.dalo.patterns.creational_patterns.factory_method.dao;
-
-import com.dalo.patterns.creational_patterns.factory_method.dao.EmployeeDao;
-import com.dalo.patterns.creational_patterns.factory_method.entity.Employee;
-
-public class PostgresEmployeeDao implements EmployeeDao {
-    @Override
-    public Employee getEmployeeById(Long id) {
-        return null;
-    }
-
-    @Override
-    public void deleteEmployee(Long id) {
-
-    }
-}
