--- src/main/java/com/dalo/patterns/creational_patterns/factory_method/entity/Employee.java
@@ -1,5 +1,0 @@
-package com.dalo.patterns.creational_patterns.factory_method.entity;
-
-public class Employee {
-    private Long id;
-}
