--- src/main/java/com/dalo/patterns/creational_patterns/pototype/model/Employee.java
@@ -1,20 +1,0 @@
-package com.dalo.patterns.creational_patterns.pototype.model;
-
-public class Employee {
-    private Long id;
-    private String firstName;
-    private String lastName;
-
-    public Employee() {
-
-    }
-    public Employee(String firstName, String lastName) {
-        this.firstName = firstName;
-        this.lastName = lastName;
-    }
-
-    public Employee copy() {
-        Employee employeeClone = new Employee(this.firstName, this.lastName);
-        return employeeClone;
-    }
-}
