--- src/main/java/com/dalo/patterns/creational_patterns/singleton/dao/impl/EmployeeDaoImpl.java
@@ -1,13 +1,0 @@
-package com.dalo.patterns.creational_patterns.singleton.dao.impl;
-
-import com.dalo.patterns.creational_patterns.pototype.model.Employee;
-import com.dalo.patterns.creational_patterns.singleton.dao.EmployeeDao;
-
-public class EmployeeDaoImpl implements EmployeeDao {
-
-    @Override
-    public Employee getEmployeeByName(String name) {
-        //implementation
-        return null;
-    }
-}
