--- src/main/java/com/dalo/patterns/creational_patterns/singleton/model/Employee.java
@@ -1,4 +1,0 @@
-package com.dalo.patterns.creational_patterns.singleton.model;
-
-public class Employee {
-}
